Clean-up namespace of the tests for tf2_sensor_msgs.
since both added and the original version are the tests for the same
function with different inputs.
