+++ tf2_sensor_msgs/CMakeLists.txt
@@ -24,7 +24,6 @@ catkin_python_setup()
 
 if(CATKIN_ENABLE_TESTING)
   catkin_add_nosetests(test/test_tf2_sensor_msgs.py)
-  catkin_add_nosetests(test/test_tf2_sensor_msgs_multichannel.py)
 
 find_package(catkin REQUIRED COMPONENTS sensor_msgs rostest tf2_ros tf2)
